pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentsh"
version = "0.1.0"
description = "AI-enhanced terminal shell with LLM-powered capabilities"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "AgentSH Team" }
]
keywords = ["ai", "shell", "terminal", "llm", "agent", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Shells",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    # PTY and terminal
    "ptyprocess>=0.7.0",
    "prompt-toolkit>=3.0",

    # Configuration and validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",

    # Logging
    "structlog>=24.0",

    # LLM providers
    "anthropic>=0.28.0",
    "openai>=1.3.0",

    # Workflow orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "mypy>=1.5",
    "black>=23.0",
    "ruff>=0.1.0",
    "bandit>=1.7",
    "pre-commit>=3.0",
    "pip-audit>=2.6.0",
    "safety>=2.3.0",
]

vector = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

ssh = [
    "paramiko>=3.0",
    "fabric>=3.0",
]

metrics = [
    "prometheus-client>=0.17.0",
]

robotics = [
    # ROS2 dependencies managed externally
]

all = [
    "agentsh[dev,vector,ssh,metrics]",
]

[project.scripts]
agentsh = "agentsh.__main__:main"

[project.urls]
Homepage = "https://github.com/agentsh/agentsh"
Documentation = "https://github.com/agentsh/agentsh/docs"
Repository = "https://github.com/agentsh/agentsh"

[tool.hatch.build.targets.wheel]
packages = ["src/agentsh"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hatch
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["agentsh"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "security: marks tests as security tests",
]

[tool.coverage.run]
source = ["src/agentsh"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70
